+++ test/libsolidity/smtCheckerTests/loops/while_loop_array_assignment_storage_storage.sol
@@ -20,9 +20,7 @@ contract LoopFor2 {
 		assert(b[0] == 900);
 	}
 }
-// ====
-// SMTSolvers: cvc4
 // ----
-// Warning 4661: (296-316): Assertion violation happens here.
-// Warning 4661: (320-339): Assertion violation happens here.
-// Warning 4661: (343-362): Assertion violation happens here.
+// Warning 1218: (229-234): Error trying to invoke SMT solver.
+// Warning 6328: (320-339): Assertion violation happens here.
+// Warning 6328: (343-362): Assertion violation happens here.