+++ test/libsolidity/smtCheckerTests/operators/delete_multid_array.sol
@@ -37,8 +37,6 @@ contract C {
 		b[x][y] = z;
 	}
 }
-// ====
-// SMTSolvers: cvc4
 // ----
-// Warning 4661: (372-392): Assertion violation happens here.
-// Warning 4661: (617-637): Assertion violation happens here.
+// Warning 6328: (372-392): Assertion violation happens here.
+// Warning 6328: (617-637): Assertion violation happens here.