+++ Changelog.md
@@ -8,9 +8,11 @@ Language Features:
 
 
 Compiler Features:
+ * Export compiler-generated utility sources via standard-json or combined-json.
+ * SMTChecker: Support ``revert()``.
  * SMTChecker: Support shifts.
  * SMTChecker: Support structs.
- * SMTChecker: Support ``type(T).min`` and ``type(T).max``.
+ * SMTChecker: Support ``type(T).min``, ``type(T).max``, and ``type(I).interfaceId``.
  * Yul Optimizer: Prune unused parameters in functions.
  * Yul Optimizer: Inline into functions further down in the call graph first.
  * Yul Optimizer: Try to simplify function names.
@@ -18,6 +20,7 @@ Compiler Features:
 
 Bugfixes:
  * Type Checker: Disallow ``virtual`` for modifiers in libraries.
+ * Type Checker: Correct the warning for homonymous, but not shadowing declarations.
  * ViewPureChecker: Prevent visibility check on constructors.
  * Type system: Fix internal error on implicit conversion of contract instance to the type of its ``super``.
  * Type system: Fix named parameters in overloaded function and event calls being matched incorrectly if the order differs from the declaration.